cbor_next::data_item

Enum DataItem

#[non_exhaustive]
pub enum DataItem {
    Unsigned(u64),
    Signed(u64),
    Byte(ByteContent),
    Text(TextContent),
    Array(ArrayContent),
    Map(MapContent),
    Tag(TagContent),
    Boolean(bool),
    Null,
    Undefined,
    Floating(f64),
    GenericSimple(SimpleValue),
}

Enum representing different types of data item that can be encoded or decoded in CBOR (Concise Binary Object Representation).

CBOR is a data format designed for small code and message size, often used in constrained environments. This DataItem enum covers all major types defined in the CBOR specification (RFC 8949).

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Unsigned(u64)

Unsigned integer represented by CBOR major type 0.

This variant can hold non-negative integer values up to u64::MAX.

Signed(u64)

Negative integer represented by CBOR major type 1.

This variant stores the absolute value minus one of the negative integer. For example, a CBOR negative integer representing -1 would store 0, -10 would store 9. The actual negative value is derived as -(1 + value).

Byte(ByteContent)

Byte string represented by CBOR major type 2.

Contains an arbitrary sequence of bytes.

Text(TextContent)

UTF-8 string (text string) represented by CBOR major type 3.

Contains a sequence of Unicode characters encoded as UTF-8.

Array(ArrayContent)

Array of CBOR data items represented by CBOR major type 4.

An ordered sequence of zero or more CBOR data items.

Map(MapContent)

Map of CBOR key-value pairs represented by CBOR major type 5.

Keys within a map must be unique

Tag(TagContent)

Tagged item (semantic tag) represented by CBOR major type 6.

Consists of an unsigned integer (the tag) and a single CBOR data item (the tagged content). Tags provide semantic information about the enclosed data item, allowing for type extension or application-specific interpretations.

Boolean(bool)

Boolean represented as a simple value within CBOR major type 7.

Can be either true or false.

Null

Null represented as a simple value within CBOR major type 7.

Represents the absence of a value.

Undefined

Undefined represented as a simple value within CBOR major type 7.

Distinct from Null, it represents an undefined state.

Floating(f64)

Floating-point number represented as a simple value within CBOR major type 7.

Can represent half-precision (16-bit), single-precision (32-bit), or double-precision (64-bit) floating-point numbers. but locally saves data as f64

GenericSimple(SimpleValue)

An generic simple value represented by CBOR major type 7.

This variant handles simple values that are not explicitly covered by Boolean, Null, Undefined, or Floating. These generic simple values have a numerical representation as defined in the CBOR specification.

Implementations

impl DataItem

pub fn is_unsigned_integer(&self) -> bool

Is a unsigned integer value?

Example

use cbor_next::DataItem;

assert!(DataItem::from(20).is_unsigned_integer());

pub fn is_signed_integer(&self) -> bool

Is a signed integer value?

Example

use cbor_next::DataItem;

assert!(DataItem::from(-32).is_signed_integer());

pub fn is_byte(&self) -> bool

Is a byte value?

Example

use cbor_next::DataItem;

assert!(DataItem::from(vec![65u8, 63, 62].as_slice()).is_byte());

pub fn is_text(&self) -> bool

Is a text value?

Example

use cbor_next::DataItem;

assert!(DataItem::from("example").is_text());

pub fn is_array(&self) -> bool

Is a array value?

Example

use cbor_next::DataItem;

assert!(DataItem::from(vec![12]).is_array());

pub fn is_map(&self) -> bool

Is a map value?

Example

use cbor_next::DataItem;
use indexmap::IndexMap;

assert!(DataItem::from(IndexMap::from_iter(vec![(12, "a")])).is_map());

pub fn is_tag(&self) -> bool

Is a tag value?

Example

use cbor_next::{DataItem, TagContent};

assert!(DataItem::from(TagContent::from((12, 20))).is_tag());

pub fn is_boolean(&self) -> bool

Is a boolean value?

Example

use cbor_next::DataItem;

assert!(DataItem::from(false).is_boolean());

pub fn is_null(&self) -> bool

Is a null value?

Example

use cbor_next::DataItem;

assert!(DataItem::Null.is_null());

pub fn is_undefined(&self) -> bool

Is a undefined value?

Example

use cbor_next::DataItem;

assert!(DataItem::Undefined.is_undefined());

pub fn is_floating(&self) -> bool

Is a floating value?

Example

use cbor_next::DataItem;

assert!(DataItem::from(3.0).is_floating());

pub fn is_simple(&self) -> bool

Is a simple value?

Example

use cbor_next::{DataItem, SimpleValue};

assert!(DataItem::from(SimpleValue::try_from(45).unwrap()).is_simple());

pub fn is_generic_simple(&self) -> bool

Is a generic simple value?

Example

use cbor_next::{DataItem, SimpleValue};

assert!(DataItem::from(SimpleValue::try_from(45).unwrap()).is_generic_simple());

pub fn check_inner(&self, checker: impl Fn(&Self) -> bool) -> bool

Recursively checks nested CBOR data items until a non-tag item is found, then applies the given checker function to that item.

This is particularly useful for examining the underlying value of tagged data items without manually unwrapping each layer of tags. Also

Example

use cbor_next::{DataItem, TagContent};

let tag = DataItem::from(TagContent::from((20, TagContent::from((30, "abc")))));
assert!(tag.check_inner(DataItem::is_text));

Note

The function will skip all outer tags before applying the checker. If you need to check the tags themselves, use DataItem::is_tag directly

pub fn as_unsigned(&self) -> Option<u64>

Get as unsigned number

Example

use cbor_next::DataItem;

assert_eq!(DataItem::from(20).as_unsigned(), Some(20));

pub fn as_signed(&self) -> Option<i128>

Get as signed number

Example

use cbor_next::DataItem;

assert_eq!(DataItem::from(-21).as_signed(), Some(-21));

pub fn as_byte(&self) -> Option<Vec<u8>>

Get as byte

Example

use cbor_next::DataItem;

assert_eq!(
    DataItem::from(vec![0x6a].as_slice()).as_byte(),
    Some(vec![0x6a])
);

pub fn as_text(&self) -> Option<String>

Get as text

Example

use cbor_next::DataItem;

assert_eq!(DataItem::from("cbor").as_text(), Some("cbor".to_string()));

pub fn as_array(&self) -> Option<&[DataItem]>

Get as array

Example

use cbor_next::DataItem;

assert_eq!(DataItem::from(vec![12u64]).as_array().unwrap(), [12.into()]);

pub fn as_map(&self) -> Option<&IndexMap<DataItem, DataItem>>

Get as map

Example

use cbor_next::DataItem;
use indexmap::IndexMap;

assert_eq!(
    DataItem::from(IndexMap::<DataItem, DataItem>::new()).as_map(),
    Some(&IndexMap::new())
);

pub fn as_tag(&self) -> Option<(u64, &DataItem)>

Get as tag

Example

use cbor_next::{DataItem, TagContent};

assert_eq!(
    DataItem::from(TagContent::from((20, -21))).as_tag(),
    Some((20, &DataItem::Signed(20)))
);

pub fn as_boolean(&self) -> Option<bool>

Get as boolean number

Example

use cbor_next::DataItem;

assert_eq!(DataItem::from(true).as_boolean(), Some(true));

pub fn as_floating(&self) -> Option<f64>

Get as floating number

Example

use cbor_next::DataItem;

assert_eq!(DataItem::from(-20.0).as_floating(), Some(-20.0));

pub fn as_simple(&self) -> Option<u8>

Get as simple index value

Example

use cbor_next::{DataItem, SimpleValue};

assert_eq!(
    DataItem::from(SimpleValue::try_from(10).unwrap()).as_simple(),
    Some(10)
);

pub fn as_inner<T>( &self, extractor: impl Fn(&Self) -> Option<T>, ) -> Option<(Vec<u64>, T)>

Recursively extract tagged values, collecting all tag numbers and returning them with the extracted value. Tags are vector of tag numbers in outer-to-inner order

When extractor is tag extractor i.e DataItem::as_tag than this would always return None since it only supports non tag extract but would successfully returns tag list

Example

use cbor_next::{DataItem, TagContent};

let tag = DataItem::from(TagContent::from((20, TagContent::from((30, -21)))));
let tag_unwrapped = tag.as_inner(DataItem::as_signed);
assert_eq!(tag_unwrapped, Some((vec![20, 30], -21)));

pub fn major_type(&self) -> u8

Get a major type of a value

pub fn encode(&self) -> Vec<u8>

Get a CBOR encoded representation of value

Example

use cbor_next::DataItem;

let value = DataItem::Unsigned(10_000_000);
let vector_data = vec![0x1a, 0x00, 0x98, 0x96, 0x80];
assert_eq!(value.encode(), vector_data);

pub fn decode(val: &[u8]) -> Result<Self, Error>

Decode a CBOR representation to a value

Example

use cbor_next::DataItem;

let vector_data = vec![0x1a, 0x00, 0x98, 0x96, 0x80];
let value = DataItem::Unsigned(10_000_000);
assert_eq!(DataItem::decode(&vector_data).unwrap(), value);

Errors

If provided bytes cannot be converted to CBOR

pub fn is_deterministic(&self, mode: &DeterministicMode) -> bool

Check current data item is deterministic form

pub fn deterministic(self, mode: &DeterministicMode) -> Self

Get a deterministic ordering form in provided mode