cbor_next::index

Trait Get

pub trait Get<Idx>
where
    Idx: Sealed,
{
    // Required methods
    fn get(&self, idx: Idx) -> Option<&Self>;
    fn get_mut(&mut self, idx: Idx) -> Option<&mut Self>;
}

Trait which is used to get a data item out of a data item.

An array is indexed by position, a map by key. A key is given either as a DataItem or, for the common case of a text key, as a string slice.

Required Methods

fn get(&self, idx: Idx) -> Option<&Self>

Get the item stored at the index, or None when the index is absent or the item is neither an array nor a map.

Example

use cbor_next::{DataItem, Get};

let array = DataItem::array([10]);
assert_eq!(array.get(0), Some(&DataItem::from(10)));
assert_eq!(array.get(2), None);

let map = DataItem::map([("abc", 10)]);
assert_eq!(map.get("abc"), Some(&DataItem::from(10)));
assert_eq!(map.get(&DataItem::from("abc")), Some(&DataItem::from(10)));
assert_eq!(map.get(&DataItem::Unsigned(11)), None);

fn get_mut(&mut self, idx: Idx) -> Option<&mut Self>

Get the item stored at the index mutably.

Example

use cbor_next::{DataItem, Get};

let mut array = DataItem::array([10]);
*array.get_mut(0).unwrap() = DataItem::from(20);
assert_eq!(array.get(0), Some(&DataItem::from(20)));

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".