cbor_next::value

Enum Value

#[non_exhaustive]
pub enum Value {
    Unsigned(u64),
    Signed(u64),
    Byte(Vec<u8>),
    Text(String),
    Array(Vec<Value>),
    Map(IndexMap<Value, Value>),
    Tag(u64, Box<Value>),
    Boolean(bool),
    Null,
    Undefined,
    Floating(f64),
    UnknownSimple(SimpleNumber),
}

Enum representing different types of values that can be encoded or decoded in CBOR (Concise Binary Object Representation).

CBOR is a data format designed for small code and message size, often used in constrained environments. This Value enum covers all major types defined in the CBOR specification (RFC 8949).

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Unsigned(u64)

Unsigned integer represented by CBOR major type 0.

This variant can hold non-negative integer values up to u64::MAX.

Signed(u64)

Negative integer represented by CBOR major type 1.

This variant stores the absolute value minus one of the negative integer. For example, a CBOR negative integer representing -1 would store 0, -10 would store 9. The actual negative value is derived as -(1 + value).

Byte(Vec<u8>)

Byte string represented by CBOR major type 2.

Contains an arbitrary sequence of bytes.

Text(String)

UTF-8 string (text string) represented by CBOR major type 3.

Contains a sequence of Unicode characters encoded as UTF-8.

Array(Vec<Value>)

Array of CBOR values represented by CBOR major type 4.

An ordered sequence of zero or more CBOR data items.

Map(IndexMap<Value, Value>)

Map of CBOR key-value pairs represented by CBOR major type 5.

Keys within a map must be unique

Tag(u64, Box<Value>)

Tagged item (semantic tag) represented by CBOR major type 6.

Consists of an unsigned integer (the tag) and a single CBOR data item (the tagged content). Tags provide semantic information about the enclosed data item, allowing for type extension or application-specific interpretations.

Boolean(bool)

Boolean value represented as a simple value within CBOR major type 7.

Can be either true or false.

Null

Null value represented as a simple value within CBOR major type 7.

Represents the absence of a value.

Undefined

Undefined value represented as a simple value within CBOR major type 7.

Distinct from Null, it represents an undefined state.

Floating(f64)

Floating-point number represented as a simple value within CBOR major type 7.

Can represent half-precision (16-bit), single-precision (32-bit), or double-precision (64-bit) floating-point numbers. but locally saves data as f64

UnknownSimple(SimpleNumber)

An unknown simple value represented by CBOR major type 7.

This variant handles simple values that are not explicitly covered by Boolean, Null, Undefined, or Floating. These unknown simple values have a numerical representation as defined in the CBOR specification.

Implementations

impl Value

pub fn is_unsigned_integer(&self) -> bool

Is a unsigned integer value?

Example

use cbor_next::Value;

assert!(Value::Unsigned(20).is_unsigned_integer());

pub fn is_signed_integer(&self) -> bool

Is a signed integer value?

Example

use cbor_next::Value;

assert!(Value::from(-32).is_signed_integer());

pub fn is_byte(&self) -> bool

Is a byte value?

Example

use cbor_next::Value;

assert!(Value::Byte(vec![65, 63, 62]).is_byte());

pub fn is_text(&self) -> bool

Is a text value?

Example

use cbor_next::Value;

assert!(Value::Text("example".to_string()).is_text());

pub fn is_array(&self) -> bool

Is a array value?

Example

use cbor_next::Value;

assert!(Value::Array(vec![12.into()]).is_array());

pub fn is_map(&self) -> bool

Is a map value?

Example

use cbor_next::Value;
use indexmap::IndexMap;

assert!(Value::Map(IndexMap::new()).is_map());

pub fn is_tag(&self) -> bool

Is a tag value?

Example

use cbor_next::Value;

assert!(Value::Tag(12, Box::new(Value::Unsigned(20))).is_tag());

pub fn is_boolean(&self) -> bool

Is a boolean value?

Example

use cbor_next::Value;

assert!(Value::Boolean(false).is_boolean());

pub fn is_null(&self) -> bool

Is a null value?

Example

use cbor_next::Value;

assert!(Value::Null.is_null());

pub fn is_undefined(&self) -> bool

Is a undefined value?

Example

use cbor_next::Value;

assert!(Value::Undefined.is_undefined());

pub fn is_floating(&self) -> bool

Is a floating value?

Example

use cbor_next::Value;

assert!(Value::Floating(3.0).is_floating());

pub fn is_simple(&self) -> bool

Is a simple value?

Example

use cbor_next::{SimpleNumber, Value};

assert!(Value::UnknownSimple(SimpleNumber::try_from(45).unwrap()).is_simple());

pub fn is_unknown_simple(&self) -> bool

Is a unknown simple value?

Example

use cbor_next::{SimpleNumber, Value};

assert!(Value::UnknownSimple(SimpleNumber::try_from(45).unwrap()).is_simple());

pub fn as_unsigned(&self) -> Option<u64>

Get as unsigned number

Example

use cbor_next::Value;

assert!(Value::Unsigned(20).as_unsigned().is_some());

pub fn as_signed(&self) -> Option<i128>

Get as signed number

Example

use cbor_next::Value;

assert!(Value::Signed(20).as_signed().is_some());

pub fn as_byte(&self) -> Option<&Vec<u8>>

Get as byte number

Example

use cbor_next::Value;

assert!(Value::Byte(vec![]).as_byte().is_some());

pub fn as_text(&self) -> Option<&str>

Get as text number

Example

use cbor_next::Value;

assert!(Value::Text(String::new()).as_text().is_some());

pub fn as_array(&self) -> Option<&Vec<Value>>

Get as array number

Example

use cbor_next::Value;

assert!(Value::Array(vec![]).as_array().is_some());

pub fn as_map(&self) -> Option<&IndexMap<Value, Value>>

Get as map number

Example

use cbor_next::Value;
use indexmap::IndexMap;

assert!(Value::Map(IndexMap::new()).as_map().is_some());

pub fn as_tag(&self) -> Option<(&u64, &Value)>

Get as tag number

Example

use cbor_next::Value;

assert!(
    Value::Tag(20, Box::new(Value::Signed(20)))
        .as_tag()
        .is_some()
);

pub fn as_boolean(&self) -> Option<bool>

Get as boolean number

Example

use cbor_next::Value;

assert!(Value::Boolean(true).as_boolean().is_some());

pub fn as_floating(&self) -> Option<f64>

Get as floating number

Example

use cbor_next::Value;

assert!(Value::Floating(-20.0).as_floating().is_some());

pub fn as_simple(&self) -> Option<u8>

Get as simple index value

Example

use cbor_next::{SimpleNumber, Value};

assert!(
    Value::UnknownSimple(SimpleNumber::try_from(10).unwrap())
        .as_simple()
        .is_some()
);

pub fn major_type(&self) -> u8

Get a major type of a value

pub fn encode(&self) -> Vec<u8>

Get a CBOR encoded representation of value

Example

use cbor_next::Value;

let value = Value::Unsigned(10_000_000);
let vector_data = vec![0x1a, 0x00, 0x98, 0x96, 0x80];
assert_eq!(value.encode(), vector_data);

pub fn decode(val: &[u8]) -> Result<Self, Error>

Decode a CBOR representation to a value

Example

use cbor_next::Value;

let vector_data = vec![0x1a, 0x00, 0x98, 0x96, 0x80];
let value = Value::Unsigned(10_000_000);
assert_eq!(Value::decode(&vector_data).unwrap(), value);

Errors

If provided bytes cannot be converted to CBOR

pub fn deterministic(self, mode: &DeterministicMode) -> Self

Get a deterministic form in provided mode