Crate cbor_next

Library to handle a Concise Binary Object Representation (CBOR)

Modules

content

Module for different type of content

data_item

Module containing a data item

deterministic

Module containing different deterministic mode

error

Module containing different type of error

index

Module for index

Structs

ArrayContent

Struct which holds a array content

ByteContent

Struct which holds a byte data

MapContent

Struct which holds a map content

SimpleValue

struct representing simple value which only allow number between 0-19 and 32-255. This struct is simple wrapper above u8 with validation

TagContent

struct which holds tag related information such as tag number and content of tag

TextContent

Struct which holds a text content

Enums

DataItem

Enum representing different types of data item that can be encoded or decoded in CBOR (Concise Binary Object Representation).

DeterministicMode

Different mode supported for deterministic format

Traits

Get

Trait which is used to get a data item from data item